backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psutil==5.9.6
python-dotenv==1.0.0
gunicorn==21.2.0
